name and material types as properties
name and material_types are now a properties for Species and all Subclasses.
